requirements.txt
aiohttp==3.7.4.post0
aiosqlite==0.17.0
APScheduler==3.7.0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
colorlog==6.4.1
discord==1.7.3
discord.py==1.7.3
hikari==2.0.0.dev102
hikari-lightbulb==1.3.0
idna==3.2
multidict==5.1.0
pure25519==0.0.1
pypiapijson==2.1
pytz==2021.1
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.2
tzlocal==2.1
urllib3==1.26.6
uvloop==0.16.0
yarl==1.6.3
